make generate passphrase as independence function, update related
