+++ templates/index.html
@@ -77,7 +77,7 @@
         });
     </script>
     <footer class="text-center mt-5">
-        <script>document.write(new Date().getFullYear());</script>, musicsms
+        <script>document.write(new Date().getFullYear());</script> - musicsms
     </footer>
 </body>
 </html>
